docs: документация для server, client и common
