+++ src/main/java/com/cipher/Launcher.java
@@ -8,10 +8,22 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.cloud.openfeign.EnableFeignClients;
 import org.springframework.context.ConfigurableApplicationContext;
 
+/**
+ * Главный класс приложения, отвечающий за запуск Spring Boot и JavaFX.
+ * Инициализирует глобальный обработчик исключений, запускает JavaFX приложение
+ * и контекст Spring в отдельных потоках.
+ */
 @SpringBootApplication
 @EnableFeignClients
 public class Launcher {
 
+    /**
+     * Точка входа в приложение.
+     * Инициализирует глобальный обработчик исключений, запускает JavaFX приложение
+     * и контекст Spring Boot.
+     *
+     * @param args аргументы командной строки
+     */
     public static void main(String[] args) {
         GlobalExceptionHandler.registerGlobalExceptionHandler();
