+++ src/main/java/com/cipher/client/handler/GlobalExceptionHandler.java
@@ -3,7 +3,21 @@ package com.cipher.client.handler;
 import javafx.application.Platform;
 import javafx.scene.control.Alert;
 
+/**
+ * Глобальный обработчик необработанных исключений.
+ * Перехватывает исключения, которые не были обработаны в других местах приложения,
+ * и показывает пользователю диалоговое окно с ошибкой перед завершением приложения.
+ * Реализует интерфейс {@link Thread.UncaughtExceptionHandler}.
+ */
 public class GlobalExceptionHandler implements Thread.UncaughtExceptionHandler {
+
+    /**
+     * Обрабатывает необработанное исключение в потоке.
+     * Показывает диалоговое окно с ошибкой и завершает приложение.
+     *
+     * @param t поток, в котором произошло исключение
+     * @param e исключение, которое было брошено
+     */
     @Override
     public void uncaughtException(Thread t, Throwable e) {
         Platform.runLater(() -> {
@@ -15,6 +29,10 @@ public class GlobalExceptionHandler implements Thread.UncaughtExceptionHandler {
         });
     }
 
+    /**
+     * Регистрирует глобальный обработчик исключений для всех потоков.
+     * Должен быть вызван при запуске приложения.
+     */
     public static void registerGlobalExceptionHandler() {
         Thread.setDefaultUncaughtExceptionHandler(new GlobalExceptionHandler());
     }
