+++ src/main/java/com/cipher/client/service/SeedService.java
@@ -1,0 +1,5 @@
+package com.cipher.client.service;
+
+public interface SeedService {
+    String generateAccount();
+}
