+++ src/main/java/com/cipher/client/service/impl/SeedServiceImpl.java
@@ -1,17 +1,17 @@
-package com.cipher.client.service;
+package com.cipher.client.service.impl;
 
 import com.cipher.client.exceptions.SeedGenerationException;
 import com.cipher.client.feign.AccountApiClient;
+import com.cipher.client.service.SeedService;
 import com.cipher.client.utils.KeysUtils;
 import com.cipher.client.utils.NetworkUtils;
+import com.cipher.common.api.AuthApi;
 import com.cipher.common.dto.AuthRequest;
 import com.cipher.common.exception.AuthException;
 import com.cipher.common.exception.CryptoException;
 import com.cipher.common.exception.NetworkException;
 import lombok.RequiredArgsConstructor;
 import org.bitcoinj.crypto.MnemonicCode;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -22,13 +22,25 @@ import java.security.*;
 import java.util.Arrays;
 import java.util.List;
 
+/**
+ * Реализация сервиса {@link SeedService} для генерации seed-фраз и создания аккаунтов.
+ * Обеспечивает создание новой seed-фразы и регистрацию аккаунта на сервере.
+ */
 @Service
 @Transactional
 @RequiredArgsConstructor
-public class SeedServiceImpl {
+public class SeedServiceImpl implements SeedService {
     private final AccountApiClient accountApiClient;
-    private static final Logger logger = LoggerFactory.getLogger(SeedServiceImpl.class);
 
+    /**
+     * Генерирует новую seed-фразу и регистрирует аккаунт на сервере.
+     *
+     * @return сгенерированная seed-фраза в виде строки из 12 слов
+     * @throws AuthException если аккаунт с таким ключом уже существует
+     * @throws NetworkException если возникли проблемы с сетью
+     * @throws CryptoException если произошла ошибка генерации ключей
+     * @throws SeedGenerationException если не удалось сгенерировать seed-фразу
+     */
     public String generateAccount() {
         NetworkUtils.checkNetworkConnection();
         List<String> words = null;
@@ -62,6 +74,12 @@ public class SeedServiceImpl {
         }
     }
 
+    /**
+     * Генерирует мнемоническую фразу из 12 слов на основе энтропии.
+     *
+     * @return список из 12 слов мнемонической фразы
+     * @throws SeedGenerationException если генерация не удалась
+     */
     private List<String> generatedWordsForSeed() {
         byte[] entropy = null;
         try {
