+++ src/main/java/com/cipher/common/dto/NonceRequest.java
@@ -1,3 +1,8 @@
 package com.cipher.common.dto;
 
+/**
+ * Запрос на получение одноразового числа (nonce).
+ *
+ * @param userId уникальный идентификатор пользователя
+ */
 public record NonceRequest(String userId) {}
