+++ src/main/java/com/cipher/common/dto/NonceResponse.java
@@ -1,3 +1,8 @@
 package com.cipher.common.dto;
 
+/**
+ * Ответ с одноразовым числом (nonce) для аутентификации.
+ *
+ * @param nonce одноразовое число в виде строки
+ */
 public record NonceResponse(String nonce) {}
