--- src/main/java/com/cipher/online/dto/ErrorResponse.java
@@ -1,13 +1,0 @@
-package com.cipher.online.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-@AllArgsConstructor
-public class ErrorResponse {
-    private String code;
-    private String message;
-}
