--- src/main/java/com/cipher/online/handler/GlobalExceptionHandler.java
@@ -1,37 +1,0 @@
-package com.cipher.online.handler;
-
-import com.cipher.common.exception.AccountAlreadyExistsException;
-import com.cipher.common.exception.SeedNotFoundException;
-import com.cipher.online.dto.ErrorResponse;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ControllerAdvice;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-
-@ControllerAdvice
-public class GlobalExceptionHandler {
-
-    @ExceptionHandler(AccountAlreadyExistsException.class)
-    public ResponseEntity<ErrorResponse> handleAccountExists(AccountAlreadyExistsException ex) {
-        return ResponseEntity.status(HttpStatus.CONFLICT)
-                .body(new ErrorResponse("ACCOUNT_EXISTS", ex.getMessage()));
-    }
-
-    @ExceptionHandler(SeedNotFoundException.class)
-    public ResponseEntity<ErrorResponse> handleSeedNotFound(SeedNotFoundException ex) {
-        return ResponseEntity.status(HttpStatus.NOT_FOUND)
-                .body(new ErrorResponse("ACCOUNT_NOT_FOUND", ex.getMessage()));
-    }
-
-    @ExceptionHandler(SecurityException.class)
-    public ResponseEntity<ErrorResponse> handleSecurityException(SecurityException ex) {
-        return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
-                .body(new ErrorResponse("SECURITY_ERROR", ex.getMessage()));
-    }
-
-    @ExceptionHandler(Exception.class)
-    public ResponseEntity<ErrorResponse> handleGenericException(Exception ex) {
-        return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
-                .body(new ErrorResponse("INTERNAL_ERROR", "Внутренняя ошибка сервера"));
-    }
-}
