--- src/main/java/com/cipher/online/repository/UserRepository.java
@@ -1,13 +1,0 @@
-package com.cipher.online.repository;
-
-import com.cipher.common.entity.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-    Optional<User> findByUserId(String userId);
-    boolean existsByUserId(String userId);
-}
