+++ src/main/java/com/cipher/server/dto/UserChangeNicknameDTO.java
@@ -1,4 +1,4 @@
-package com.cipher.online.dto;
+package com.cipher.server.dto;
 
 import lombok.Getter;
 import lombok.Setter;
